requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
jinja2==3.1.2
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6
google-generativeai==0.3.0
gTTS==2.3.2
pygame==2.5.2
yfinance==0.2.28